refactor(sync): move the embedded-model command into sync_autopilot (#440 phase 3)
Decision (B) made structural: lift the legacy all-in-one command (`slides_sync_cmd`)
and the four model-client constructors (`_resolve_judge` / `_resolve_recoverer` /
`_resolve_verifier` + the translator) — the only code that references the OpenRouter/
Ollama client classes — out of the agent-path `sync` module into a new sibling
`sync_autopilot` module. The verb group registers `autopilot` LAZILY (LazyGroup spec),
so importing `clm.cli.commands.slides.sync` — which happens on every `clm slides`
invocation — pulls in neither the OpenAI SDK nor the autopilot module.

Net: the agent verbs (report / verify / task / accept / apply) live in a module that
binds no model-client class, and "no live LLM reachable on the agent path / in CI" is
a property of the import graph, not a discipline. No user-facing change — verbs, flags
and outputs are identical; `autopilot` (and only it) loads the clients on first use.

Mechanics:
- sync.py keeps every model-free helper (path resolution, exit codes, output, the
  batch sweep, _run_report/_run_verify/_run_rebaseline, the stale/cold hints, the
  progress wrappers); it imports only `has_openrouter_api_key` (an env check) from a
  client module. sync_autopilot imports the shared helpers from sync.py.
- A PEP 562 `__getattr__` re-exports `slides_sync_cmd` from sync for back-compat
  (existing tests + the lazy spec) WITHOUT a module-level import of autopilot.
- translate.py (a model command of its own) lazily imports `_resolve_judge` from
  sync_autopilot so loading it does not pull autopilot onto the agent path.
- Test migration: the autopilot-command tests (test_slides_sync, test_sync_code_e2e)
  patch model factories at sync_autopilot now; the one batch test patches
  build_sync_plan at sync (the batch path lives there); translate tests patch
  `_resolve_judge` at its new home.

Adds tests/cli/test_sync_import_cleanliness.py (subprocess probes: importing sync
loads neither openai nor autopilot; resolving the autopilot command still doesn't
load the SDK). Full sync/translate/slides sweep green (2205 passed).
