+++ changelog.d/440-sync-model-import-split.changed.md
@@ -1,0 +1,1 @@
+- **The `clm slides sync` agent verbs are now import-clean of the model clients (epic #440, phase 3; decision B made structural).** The four embedded model clients (the edit judge, the new-slide translator, the cold-pair correspondence verifier, and the `--llm-recover` id recoverer) and the legacy all-in-one command that drives them moved out of the agent-path module into a sibling `sync_autopilot` module, registered on the group **lazily**. The result: importing the module the agent verbs (`report` / `verify` / `task` / `accept` / `apply`) live in — which happens on every `clm slides` invocation — pulls in **neither the OpenAI SDK nor the autopilot command**, so "no live LLM is reachable from the agent path / in CI" is now a property of the import graph rather than a discipline. The `autopilot` verb (and only it) loads the clients, on first use. No user-facing surface changes — the verbs, flags, and outputs are identical. (#366, #440)
